+++ source/dsp/inference/InferenceThread.cpp
@@ -20,7 +20,7 @@ void InferenceThread::run() {
         globalSemaphore.acquire();
         for (const auto& session : sessions) {
             if (session->sendSemaphore.try_acquire()) {
-                for (int i = 0; i < session->inferenceQueue.size(); ++i) {
+                for (size_t i = 0; i < session->inferenceQueue.size(); ++i) {
                     if (session->inferenceQueue[i].ready.try_acquire()) {
                         inference(session->currentBackend, session->inferenceQueue[i].processedModelInput, session->inferenceQueue[i].rawModelOutputBuffer);
                         session->inferenceQueue[i].done.release();
