+++ source/dsp/utils/RingBuffer.h
@@ -8,17 +8,18 @@ class RingBuffer
 public:
     RingBuffer();
 
-    void initialise(int numChannels, int numSamples);
+    void initialise(size_t numChannels, size_t numSamples);
     void reset();
-    void pushSample(float sample, int channel);
-    float popSample(int channel);
-    float getSample(int channel, unsigned int offset);
-    int getAvailableSamples(int channel);
+    void pushSample(float sample, size_t channel);
+    float popSample(size_t channel);
+    float getSample(size_t channel, size_t offset);
+    size_t getAvailableSamples(size_t channel);
 
 private:
     juce::AudioBuffer<float> buffer;
-    std::vector<int> readPos, writePos;
+    std::vector<size_t> readPos, writePos;
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (RingBuffer)
 };
+
 #endif //VAESYNTH_RINGBUFFER_H
